Move HTTP endpoints out of Program.cs
